requirements.txt
flask==2.0.1
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
pillow==8.3.1
opencv-python==4.5.3.56
gunicorn==20.1.0
dill==0.3.4
xgboost>=1.6.0
catboost>=1.0.0
-e .